Share some code between vertx-http-client 3 & 4
