+++ instrumentation/vertx-http-client/vertx-http-client-4.0/javaagent/build.gradle.kts
@@ -14,5 +14,7 @@ muzzle {
 dependencies {
   library("io.vertx:vertx-core:4.0.0")
 
+  implementation(project(":instrumentation:vertx-http-client:vertx-http-client-common:javaagent"))
+
   testInstrumentation(project(":instrumentation:netty:netty-4.1:javaagent"))
 }
