--- instrumentation/vertx-http-client/vertx-http-client-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/v4_0/client/Contexts.java
@@ -1,18 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.vertx.v4_0.client;
-
-import io.opentelemetry.context.Context;
-
-public class Contexts {
-  public final Context parentContext;
-  public final Context context;
-
-  public Contexts(Context parentContext, Context context) {
-    this.parentContext = parentContext;
-    this.context = context;
-  }
-}
