+++ instrumentation/vertx-http-client/vertx-http-client-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/v4_0/client/VertxClientTracer.java
@@ -5,25 +5,17 @@
 
 package io.opentelemetry.javaagent.instrumentation.vertx.v4_0.client;
 
-import io.opentelemetry.context.propagation.TextMapSetter;
-import io.opentelemetry.instrumentation.api.tracer.HttpClientTracer;
-import io.opentelemetry.instrumentation.api.tracer.net.NetPeerAttributes;
+import io.opentelemetry.javaagent.instrumentation.vertx.client.AbstractVertxClientTracer;
 import io.vertx.core.http.HttpClientRequest;
-import io.vertx.core.http.HttpClientResponse;
 import java.net.URI;
 import java.net.URISyntaxException;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-public class VertxClientTracer
-    extends HttpClientTracer<HttpClientRequest, HttpClientRequest, HttpClientResponse> {
+public class VertxClientTracer extends AbstractVertxClientTracer {
   private static final VertxClientTracer TRACER = new VertxClientTracer();
 
   public static VertxClientTracer tracer() {
     return TRACER;
-  }
-
-  public VertxClientTracer() {
-    super(NetPeerAttributes.INSTANCE);
   }
 
   @Override
@@ -40,39 +32,5 @@ public class VertxClientTracer
   @Nullable
   protected URI url(HttpClientRequest request) throws URISyntaxException {
     return new URI(request.absoluteURI());
-  }
-
-  @Override
-  @Nullable
-  protected Integer status(HttpClientResponse response) {
-    return response.statusCode();
-  }
-
-  @Override
-  @Nullable
-  protected String requestHeader(HttpClientRequest request, String name) {
-    return request.headers().get(name);
-  }
-
-  @Override
-  @Nullable
-  protected String responseHeader(HttpClientResponse response, String name) {
-    return response.getHeader(name);
-  }
-
-  @Override
-  protected TextMapSetter<HttpClientRequest> getSetter() {
-    return Propagator.INSTANCE;
-  }
-
-  private static class Propagator implements TextMapSetter<HttpClientRequest> {
-    private static final Propagator INSTANCE = new Propagator();
-
-    @Override
-    public void set(HttpClientRequest carrier, String key, String value) {
-      if (carrier != null) {
-        carrier.putHeader(key, value);
-      }
-    }
   }
 }
