+++ instrumentation/vertx-http-client/vertx-http-client-common/javaagent/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+  id("otel.javaagent-instrumentation")
+}
+
+dependencies {
+  compileOnly("io.vertx:vertx-core:3.0.0")
+}
