+++ instrumentation/vertx-http-client/vertx-http-client-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/client/Contexts.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.vertx.v3_0.client;
+package io.opentelemetry.javaagent.instrumentation.vertx.client;
 
 import io.opentelemetry.context.Context;
