+++ settings.gradle.kts
@@ -313,6 +313,7 @@ include(":instrumentation:vaadin-14.2:javaagent")
 include(":instrumentation:vaadin-14.2:testing")
 include(":instrumentation:vertx-http-client:vertx-http-client-3.0:javaagent")
 include(":instrumentation:vertx-http-client:vertx-http-client-4.0:javaagent")
+include(":instrumentation:vertx-http-client:vertx-http-client-common:javaagent")
 include(":instrumentation:vertx-reactive-3.5:javaagent")
 include(":instrumentation:vertx-web-3.0:javaagent")
 include(":instrumentation:vertx-web-3.0:testing")
